Moving enums to util.py
