+++ krux_boto/util.py
@@ -15,6 +15,7 @@ import string
 #
 
 import boto.utils
+from enum import Enum
 
 #
 # Internal libraries
@@ -39,3 +40,29 @@ def get_instance_region():
         get_logger('krux_boto').warn('get_instance_region failed to get the local instance region')
         raise Error('get_instance_region failed to get the local instance region')
     return zone.rstrip(string.lowercase)
+
+
+# Region codes
+class RegionCode(Enum):
+    ASH = ('ASH', 'us-east-1')
+    PDX = ('PDX', 'us-west-2')
+    DUB = ('DUB', 'eu-west-1')
+    SIN = ('SIN', 'ap-southeast-1')
+
+    def __init__(self, code, region):
+        self.code = code
+        self.region = region
+
+    # GOTCHA: It would be so nice if I can override __getitem__ function
+    # for this. However that is not possible with the current implementation.
+    # Thus, creating a class method for this.
+    @classmethod
+    def from_region(cls, region):
+        # Lower the
+        region = region.lower()
+
+        for enum in list(cls):
+            if enum.region == region:
+                return enum
+
+        raise KeyError(region)
